requirements.txt
keras~=2.13.1rc0
requests~=2.31.0
numpy~=1.24.3
Flask~=2.3.2
tensorflow~=2.13.0rc1
Pillow~=9.5.0

